requirements.txt
python-telegram-bot==21.6
openai==1.3.0
APScheduler==3.10.4
python-dotenv==1.0.0
tzdata==2024.1
httpx==0.27.2
gspread==6.1.4
google-auth==2.35.0

